+++ libreborn/include/libreborn/util.h
@@ -45,6 +45,11 @@ int is_progress_difference_significant(int32_t new_val, int32_t old_val);
 int lock_file(const char *file);
 void unlock_file(const char *file, int fd);
 
+// Access Configuration At Runtime
+const char *reborn_get_version();
+int reborn_is_headless();
+int reborn_is_server();
+
 #ifdef __cplusplus
 }
 #endif